# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DB project server.",
        "title": "DB project API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:5000",
    "basePath": "/api",
    "paths": {
...
        "/service/clear": {
            "post": {
                "description": "Clears Service info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service"
                ],
                "summary": "Clears Service info",
                "operationId": "ServiceClear",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error - Request is valid but operation failed at server side",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/service/status": {
            "get": {
                "description": "Gets Service info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service"
                ],
                "summary": "Gets Service info",
                "operationId": "GetServiceStatus",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Request is valid but operation failed at server side",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Error": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
...
        "model.Status": {
            "type": "object",
            "properties": {
                "forum": {
                    "type": "integer"
                },
                "post": {
                    "type": "integer"
                },
                "thread": {
                    "type": "integer"
                },
                "user": {
                    "type": "integer"
                }
            }
        },
...
    }
}